refaktoring/modal
Я не переносив поки що функції з modal.js Це мій основний файл який підключено в index.js Я на сьогодні все, якщо є зайваження я завтра продовжу. Функція яка вішає слухач в середині modal-markup знаходиться в modal.js  я написав над нею комент
