+++ src/js/utils/constants.js
@@ -1,0 +1,4 @@
+export const API_KEY = 'a8de9bbb748883055cd7737934b96801';
+export const BASE_URL = 'https://api.themoviedb.org/3';
+
+export const MEDIA_TYPE = 'movie';
