+++ src/js/utils/references.js
@@ -1,2 +1,4 @@
 export const addToWatchedEl = document.querySelector('.watched__btn');
 export const addToQueueEl = document.querySelector('.queue__btn');
+export const openModalFilm = document.querySelector('.main-gallery__list');
+export const modalFilmContainer = document.querySelector('.container-js');
